feat(http-request-form.tsx): Separate form into multiple components for each control
